refactored species again and calculateDwm - new class DwmCalculator
put the species specific abp equations back as something that one can
get through a method in a given species. While it seemed to make more
sense to put these in the calculator, this involved switch statements
which did not ensure that every new species would have an equation. The
lack of safety prompted me to put it back in the species class. So
AbpCalculator is gone.

DwmCalcluator is a new class however in which I have put
calculateDwm(Quadrat) instead of having it in the Quadrat class. I think
this makes more sense and I will be overloading the method to calculate
for a Stand as well.